Split scanner
