+++ corpus/latex/dashundergaps/l3doc-TUB-cls.txtt
@@ -1,0 +1,13 @@
+================================================================================
+l3doc-TUB class triggers pipe and double quote verb delimiter
+================================================================================
+\documentclass{l3doc-TUB}
+|foo|
+"bar"
+--------------------------------------------------------------------------------
+(document
+  (use
+    (cs)
+    (group (l) (name) (r)))
+  (short_verb (verb_delim) (verbatim) (verb_delim))
+  (short_verb (verb_delim) (verbatim) (verb_delim)))
