+++ corpus/latex/nlctdoc/nlctdoc-cls.txtt
@@ -1,0 +1,11 @@
+================================================================================
+nlctdoc class triggers pipe verb delimiter
+================================================================================
+\documentclass{nlctdoc}
+|foo|
+--------------------------------------------------------------------------------
+(document
+  (use
+    (cs)
+    (group (l) (name) (r)))
+  (short_verb (verb_delim) (verbatim) (verb_delim)))
