+++ corpus/platex/base/plnews-cls.txtt
@@ -1,0 +1,11 @@
+================================================================================
+olnews class triggers pipe verb delimiter
+================================================================================
+\documentclass{plnews}
+|foo|
+--------------------------------------------------------------------------------
+(document
+  (use
+    (cs)
+    (group (l) (name) (r)))
+  (short_verb (verb_delim) (verbatim) (verb_delim)))
