+++ grammar.js
@@ -94,7 +94,7 @@ let g = {
     $.cs_ensuremath,
     $.cs_expandafter,
     $.cs_fref,
-    $.cs_gls,
+    $.cs_gls_acr,
     $.cs_glsdisp,
     $.cs_href,
     $.cs_hyperbaseurl,
@@ -172,6 +172,7 @@ let g = {
     $.octal,
     $.par,
     $.parameter_ref,
+    $.plus,
     $.r,
     $.rbrack,
     $.short_verb_delim,
